chore: add callback form for serverVersion
