+++ packages/pg-gateway/src/connection.types.ts
@@ -13,9 +13,18 @@ export type TlsInfo = {
   sniServerName?: string;
 };
 
+export const ServerStep = {
+  AwaitingInitialMessage: 'AwaitingInitialMessage',
+  PerformingAuthentication: 'PerformingAuthentication',
+  ReadyForQuery: 'ReadyForQuery',
+} as const;
+
+export type ServerStep = (typeof ServerStep)[keyof typeof ServerStep];
+
 export type ConnectionState = {
   hasStarted: boolean;
   isAuthenticated: boolean;
   clientInfo?: ClientInfo;
   tlsInfo?: TlsInfo;
+  step: ServerStep;
 };
